+++ packages/modul-components/src/components/phonefield/phonefield.html
@@ -8,20 +8,19 @@
                 'm--has-validation-message': hasValidationMessage } ]">
     <div class="m-phonefield__wrapper">
 
-        <m-select class="m-phonefield__country"
-                  :value="countryModel"
-                  :options="isoCountries"
-                  :disabled="isDisabled || isWaiting"
-                  :readonly="isReadonly"
-                  :error="hasPhonefieldError"
-                  :valid="isPhonefieldValid"
-                  :label-up="true"
-                  :label="i18nCountryLabel"
-                  width="auto"
-                  :virtual-scroll="true"
-                  @input="onContryChanged($event)"
-                  @focus="onSelectFocus"
-                  @blur="onSelectBlur">
+        <m-select-virtual-scroll class="m-phonefield__country"
+                                 :value="countryModel"
+                                 :options="isoCountries"
+                                 :disabled="isDisabled || isWaiting"
+                                 :readonly="isReadonly"
+                                 :error="hasPhonefieldError"
+                                 :valid="isPhonefieldValid"
+                                 :label-up="true"
+                                 :label="i18nCountryLabel"
+                                 width="auto"
+                                 @input="onContryChanged($event)"
+                                 @focus="onSelectFocus"
+                                 @blur="onSelectBlur">
             <div class="m-phonefield__flag-wrapper">
                 <svg class="m-phonefield__flag"
                      width="27px"
@@ -60,7 +59,7 @@
                 </m-select-item>
             </template>
 
-        </m-select>
+        </m-select-virtual-scroll>
 
         <m-input-style class="m-phonefield__number"
                        :disabled="isDisabled"
