+++ packages/modul-components/src/components/select-virtual-scroll/select-virtual-scroll.html
@@ -4,7 +4,7 @@
                :selected-items="selectedItems"
                :open.sync="open"
                :input-max-width="inputMaxWidth"
-               :virtual-scroll="virtualScroll"
+               :virtual-scroll="true"
                @open="onOpen"
                @close="onClose"
                @select-item="onSelect">
@@ -89,19 +89,27 @@
         </div>
     </template>
 
-    <template v-slot:outer-items="{item, index, props, handlers}">
-        <slot name="outer-items"
-              :index="index"
-              :item="item"
-              :props="props"
-              :handlers="handlers">
-            <template v-slot:items="{item , index }">
 
-                <slot name="items"
+    <template #outer-items="{items, getItemProps, getItemHandlers}">
+        <virtual-list :size="virtualScrollItemHeight"
+                      :remain="4"
+                      wclass="m-base-select__virtuallist">
+            <template v-for="(item, index) of items">
+                <slot name="outer-items"
                       :item="item"
-                      :index="index"><span v-html="item"></span></slot>
+                      :index="index"
+                      :props="getItemProps(item, index)"
+                      :handlers="getItemHandlers(item, index)">
+                    <m-select-item :key="index"
+                                   v-bind="getItemProps(item, index)"
+                                   v-on="getItemHandlers(item, index)">
+                        <slot name="items"
+                              :item="item"
+                              :index="index"><span v-html="item"></span></slot>
+                    </m-select-item>
+                </slot>
             </template>
-        </slot>
+        </virtual-list>
     </template>
 
 </m-base-select>
