+++ packages/modul-components/src/components/select-virtual-scroll/select-virtual-scroll.scss
@@ -1,0 +1,56 @@
+@import "commons";
+
+.m-select {
+    display: inline-flex;
+    flex-direction: column;
+
+    &:focus {
+        outline: 0 solid transparent;
+    }
+
+    &.m--has-validation-message {
+        .m-select__validation-message {
+            margin-top: $m-spacing--xs;
+        }
+    }
+
+    &:not(.m--is-disabled) {
+        .m-select__arrow {
+            cursor: pointer;
+        }
+    }
+
+    &__arrow {
+        cursor: pointer;
+        height: $m-touch-size;
+        width: calc(#{$m-touch-size} - #{$m-spacing});
+        display: flex;
+        align-items: center;
+        justify-content: left;
+
+        &__icon {
+            transition: transform $m-transition-duration ease, color $m-transition-duration ease !important;
+        }
+
+        &.m--is-open {
+            .m-select__arrow__icon {
+                transform: rotate(-180deg);
+            }
+        }
+    }
+
+    &.m--is-disabled,
+    &.m--is-readonly {
+        .m-select__arrow {
+            cursor: default;
+        }
+
+        .m-select__arrow__icon {
+            color: $m-input-style--color-disabled;
+        }
+    }
+
+    &__validation-message {
+        transition: margin-top $m-transition-duration ease;
+    }
+}
