+++ packages/modul-components/src/components/select/base-select/base-select.ts
@@ -1,5 +1,4 @@
 import { Component, Emit, Prop, Watch } from 'vue-property-decorator';
-import virtualList from 'vue-virtual-scroll-list';
 import { MediaQueries, MediaQueriesMixin } from '../../../mixins/media-queries/media-queries';
 import { ModulVue } from '../../../utils/vue/vue';
 import { MPopup } from '../../popup/popup';
@@ -7,12 +6,12 @@ import { MSelectItem } from '../../select/select-item/select-item';
 import WithRender from './base-select.html';
 import './base-select.scss';
 
+
 const DROPDOWN_STYLE_TRANSITION: string = 'max-height 0.3s ease';
 @WithRender
 @Component({
     components: {
-        MSelectItem,
-        'virtual-list': virtualList
+        MSelectItem
     },
     mixins: [
         MediaQueries
@@ -56,9 +55,6 @@ export class MBaseSelect extends ModulVue {
 
     @Prop({ default: false })
     public virtualScroll: boolean;
-
-    @Prop({ default: 52 }) // 208px / 4 -> base-select.scss
-    public virtualScrollItemHeight: string;
 
     public $refs: {
         items: HTMLUListElement;
