+++ src/storybook/src/modul-components/components/select/select.stories.ts
@@ -1,7 +1,6 @@
 import { actions } from '@storybook/addon-actions';
 import { boolean, text } from '@storybook/addon-knobs';
 import { SELECT_NAME } from '@ulaval/modul-components/dist/components/component-names';
-import { MSelectItem } from '@ulaval/modul-components/dist/components/select/select-item/select-item';
 import { modulComponentsHierarchyRootSeparator } from '../../../utils';
 
 const OPTIONS: string[] = ['apple', 'banana', 'patate', 'tomato', 'avocados', 'etc'];
@@ -198,32 +197,4 @@ export const disabledItemSelectedWithLabelClearable = () => ({
         options: OPTIONS
     }),
     template: `<m-select :options="options" :disabled="true" label="Fruits" :clearable="true" v-model="model" ></m-select>`
-});
-
-export const virtualScroll = () => ({
-    data: () => ({
-        options: buildLongList()
-    }),
-    template: `<m-select :options="options" label="Longlist" :virtual-scroll="true" ></m-select>`
-});
-
-export const withOuterItemsSlot = () => ({
-    data: () => ({
-        model: 'banane',
-        options: OPTIONS
-    }),
-    components: {
-        MSelectItem: MSelectItem
-    },
-    template: `<m-select :options="options" v-model="model">
-                <template #outer-items="{item, index, props, handlers}">
-                    <m-select-item v-if="item !== 'patate'" v-bind="props" v-on="handlers">
-                        {{ item  }}
-                    </m-select-item>
-                    <m-select-item v-else
-                                    :disabled="true">
-                        {{ item  }}
-                    </m-select-item>
-                </template>
-            </m-select>`
 });
